Add get/set interface to StatTracker
Add a set of methods to setting and returning the label value, and the
min/max and current values of the spinbox. Add a set of tests to go
with it using the pytest-qt plugin for pytest. Update requirements.txt
accordingly. Add workspace vscode settings for autocompletion of methods
within the SethTracker folder.