Add basic Gui structure
Create a central layout in seth_tracker and place StatTrackers with
the chosen attributes into it. The Trackers are first organised by
category then placed into QGroupBoxes for visual differentiation.
These group boxes are added to the central layout within a QVBox.

In its current form, the app can be used with the values hardcoded into
the StatTrackers but further work is required for saving the values and
loading them back into the app on start up.